requirements.txt
sentence_transformers
faiss-cpu
streamlit
pdfplumber
transformers
spacy
